requirements.txt
nextcord
requests
gitpython
mysql-connector-python